Update evo_trainer.py and ce_cmds_2.json: Change mutation rate parameter type from int to float in evo_trainer.py for DE optimizer. Remove redundant command entries in ce_cmds_2.json to streamline command configurations, enhancing clarity and usability for evolutionary strategies.

+++ evo_trainer.py
@@ -714,7 +714,7 @@ if __name__ == "__main__":
                        help='Initial bound for EA optimizer')
     parser.add_argument('--de_cr', type=float, default=None,
                        help='Crossover rate for DE optimizer')
-    parser.add_argument('--de_mr', type=int, default=None,
+    parser.add_argument('--de_mr', type=float, default=None,
                        help='Mutation rate for DE optimizer')
     parser.add_argument('--ga_cr', type=float, default=None,
                        help='Crossover rate for GA optimizer')
